[Major] adding Msl wrapper for localization item/dialog and moving files related
